chore: merge RateHandler tests
